requirements.txt
pandas==1.3.5
pandas_profiling==3.4.0
Pillow==9.2.0
streamlit==1.13.0
streamlit_elements==0.1.0
streamlit_option_menu==0.3.2
streamlit_pandas_profiling==0.1.3
